poornasree_ai_api/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pytz==2023.3

# Database Support
sqlalchemy==2.0.23
aiomysql==0.2.0
pymysql==1.1.0
alembic==1.13.1

# CORS Support for Flutter
starlette==0.27.0

# Advanced AI/ML Libraries for Production
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2
tokenizers==0.15.0
scikit-learn==1.3.2
nltk==3.8.1

# Document Processing (Free)
PyPDF2==3.0.1
python-docx==1.1.0
docx2txt==0.8
openpyxl==3.1.2
xlrd==2.0.1
pandas==2.1.4

# Vector Database and Search (Free - Local)
chromadb==0.4.18
numpy==1.24.3
faiss-cpu==1.7.4

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# JSON Processing
orjson==3.9.10

# Date/Time utilities
python-dateutil==2.8.2

# File handling
aiofiles==23.2.1

# Logging
loguru==0.7.2

# System monitoring
psutil==5.9.6
